+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### 5.2.3-SNAPSHOT
 - Fix occasional chunk missing light data when its original chunk was loaded and immediately unloaded within 1-10 frames (#290)
+- Fix newly loaded real chunks sometimes flickering black for a few frames without Sodium or Starlight (#290)
 
 ### 5.2.2
 - Update to Minecraft 1.21